Program testing, minor corrections and post-processing
Made all the unitary tests work, as well as provide xml for the MDA integration test that resembles approximatley the Cirrus SR22 and the Beechcraft Duchess (76). Minor changes ot the computation of the propeller depth, the wing and fuselage weight, mlg cg. Also added three functions for the post-processing of load analysis.