Aerodynamics changes
Remade the options systems for the high lift computation so that options available for vlm match as much as possible the one available for VLM. Added a new method to compute the mach interpolation curve for time saving reasons. Added a way to compute slipstream effects for low speed. Change the high lift aerodynamics computation to include the full curves proposed by Roskam for the Delta_Cl_max. Minor bug fixes as well as useless code deletion