VTP area sizing and Xfoil pre-processing
Created a Notebook that allows the user to run Xfoil beforehand and extracts the runs that gives proper results. Also changed the sizing of the VTP to allow some more use case, namely maintaining straight flight path with engine out and crosswind landing. Also added a method to compute the maximum rated power of an engine in altitude.